backend/pyproject.toml
[tool.poetry]
name = "schulte-backend"
version = "1.0.0"
description = "FastAPI backend for Schulte Table Training App"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.34.0" }
sqlalchemy = { extras = ["asyncio"], version = "^2.0.36" }
asyncpg = "^0.30.0"
alembic = "^1.14.0"
pydantic = "^2.10.0"
pydantic-settings = "^2.7.0"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
httpx = "^0.28.0"
structlog = "^24.4.0"
boto3 = "^1.36.0"
email-validator = "^2.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.25.0"
pytest-cov = "^6.0.0"
ruff = "^0.9.0"
mypy = "^1.14.0"
types-boto3 = "^1.36.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
